LayoutTests/editing/smart-lists/ordered-list-starting-from-non-one-sets-start-attribute.html

<!DOCTYPE html><!-- webkit-test-runner [ SmartListsAvailable=true ] -->
<html>
<head>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
</head>
<body>
<div id="test" contenteditable></div>
<script>
Markup.description('Typing an ordered list whose first item starts at a number greater than 1 should produce an ol element whose start attribute matches the first number typed.');

var test = document.getElementById('test');
test.focus();
[...'5. A\n6. B'].forEach(typeCharacterCommand);
Markup.dump('test');
</script>
</body>
</html>
